+++ src/main/java/gift/exception/product/ProductAlreadyExistsException.java
@@ -1,4 +1,4 @@
-package gift.exception;
+package gift.exception.product;
 
 public class ProductAlreadyExistsException extends RuntimeException {
     public ProductAlreadyExistsException(String message) {
