+++ src/main/java/gift/exception/product/ProductNotFoundException.java
@@ -1,4 +1,4 @@
-package gift.exception;
+package gift.exception.product;
 
 public class ProductNotFoundException extends RuntimeException {
     public ProductNotFoundException(String message) {
