+++ src/main/java/gift/service/AuthService.java
@@ -3,8 +3,8 @@ package gift.service;
 import gift.dto.LoginRequest;
 import gift.dto.LoginResponse;
 import gift.dto.UserRequest;
-import gift.exception.UserAlreadyExistsException;
-import gift.exception.UserNotFoundException;
+import gift.exception.user.UserAlreadyExistsException;
+import gift.exception.user.UserNotFoundException;
 import gift.jwt.JwtUtil;
 import gift.model.User;
 import gift.repository.UserDao;
